requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
httpx==0.28.1
google-genai==1.68.0
pydantic==2.12.5
